Set up framework for each device.
I'm imagining modules to encapsule each bit of code that get pulled in to build a project. All boards share the same capabilities as the WiPy. With that, WiPyFunctions can become PyComFunctions. New modules become FiPyFunctions and LoPyFunctions. Set up a server and host it has to be a first class feature.